+++ forum/src/main/java/com/solera/forum/services/UserService.java
@@ -1,11 +1,15 @@
 package com.solera.forum.services;
 
+import com.google.gson.Gson;
 import com.solera.forum.repositories.UserRepository;
 import com.solera.forum.models.User;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserService {
@@ -13,19 +17,48 @@ public class UserService {
     @Autowired
     UserRepository userRepository;
 
-    public User createUser(User user) {
-        return userRepository.save(user);
+    public ResponseEntity createUser(User user) {
+        Gson gson = new Gson();
+        if(emailAlreadyExists((user.getEmail())))  {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body("{\"msg\":\"There is already an account associated with this email.\"}");
+        }
+        User newUser = userRepository.save(user);
+        return ResponseEntity
+                .status(HttpStatus.CREATED)
+                .body("{\"msg\":\"User created successfully.\"}");
     }
 
     public List<User> getUsers() {
         return userRepository.findAll();
     }
 
-    public User getUser(int id) {
-        return userRepository.findById(id).orElseThrow(() -> new RuntimeException("User not found"));
+    public ResponseEntity getUser(int id) {
+        Optional<User> user = userRepository.findById(id);
+        if(user.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("{\"msg\":\"User not found.\"}");
+        }
+        return ResponseEntity.status(HttpStatus.OK).body(user.get());
     }
 
-    public void deleteUser(int id) {
+    public ResponseEntity deleteUser(int id) {
+        Optional<User> user = userRepository.findById(id);
+        if(user.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("{\"msg\":\"User not found.\"}");
+        }
         userRepository.deleteById(id);
+        return ResponseEntity
+                .status(HttpStatus.OK)
+                .body("{\"msg\":\"User deleted successfully.\"}");
+    }
+
+    public boolean emailAlreadyExists(String email) {
+        List<User> users = getUsers();
+        for(User user : users){
+            if(user.getEmail().equals(email)) {
+                return true;
+            }
+        }
+        return false;
     }
 }
